Move stats stuff back to the event package
